+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+[Added]
+- Added details with statistics for individually SPs and IdPs
 
 ## [v1.5.0]
 [Added]